v3.5.0
-- New Stuff

* Add "tracking changes" functionality in various places
* Add "Send to New Hex editor" action
* Add Json color the the view
* New Make Your Own proxy (MYO)
* "Search" colors are now configurables
* New Shared memory listener (Sys V and Posix)
* Save/restore JsonView state
* Manual Edit Expand all and Collapse all Json

-- Various

* Markings improvements
* Tiny selection bug
* Fix integer to double conversion for jsonview
* Fix crash bug with view tabs
* Prevent overwriting save with empty file
* Set default visual for distorm to true
* Improving Search Widget animation
* Add QScintilla in the thrid party list
* Doc & colors (matching sys themes)
* Various build improvements
* Refactoring

+++ defaultplugins/baseplugins/ciscosecret7.cpp
@@ -50,7 +50,7 @@ void CiscoSecret7::transform(const QByteArray &input, QByteArray &output)
             return;
         }
         bool k = true;
-        seed = encrypted.mid(0,2).toInt(&k);
+        seed = static_cast<uint>(encrypted.mid(0,2).toInt(&k));
         if (!k || seed > MAXSEED) {
             emit error(tr("Invalid seed, it must be an unsigned integer <= %1").arg(MAXSEED),id);
             return;
@@ -96,6 +96,6 @@ bool CiscoSecret7::setSeed(unsigned int nseed)
 void CiscoSecret7::cipher(unsigned int seed, const QByteArray &data, QByteArray &output)
 {
     for (int i = 0; i < data.size(); i ++) {
-        output.append(data.at(i) ^ ckey.at((seed + i) % ckey.size()));
+        output.append(data.at(i) ^ ckey.at(static_cast<int>(seed + static_cast<unsigned int>(i)) % ckey.size()));
     }
 }
