+++ defaultplugins/baseplugins/randomcase.cpp
@@ -48,7 +48,7 @@ void RandomCase::reRandomize()
 
 void RandomCase::transform(const QByteArray &input, QByteArray &output) {
     output.clear();
-    qsrand(QTime::currentTime().msec()); // non cryptographic quality, boo
+    qsrand(static_cast<uint>(QTime::currentTime().msec())); // non cryptographic quality, boo
     QByteArray temp;
     for (int i = 0; i < input.size(); i++) {
         temp.append(input.at(i));
